backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
openai>=1.0.0
chromadb>=0.4.0
gitpython>=3.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-multipart>=0.0.5
aiofiles>=23.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
tiktoken>=0.5.0
